feat: refactor script for package
