+++ packages/experiment-tag/src/util/url.ts
@@ -1,5 +1,4 @@
-import { EvaluationVariant, getGlobalScope } from '@amplitude/experiment-core';
-import { Variant } from '@amplitude/experiment-js-client';
+import { getGlobalScope } from '@amplitude/experiment-core';
 
 export const getUrlParams = (): Record<string, string> => {
   const globalScope = getGlobalScope();
@@ -32,31 +31,6 @@ export const removeQueryParams = (
   return urlObj.toString();
 };
 
-export const UUID = function (a?: any): string {
-  return a // if the placeholder was passed, return
-    ? // a random number from 0 to 15
-      (
-        a ^ // unless b is 8,
-        ((Math.random() * // in which case
-          16) >> // a random number from
-          (a / 4))
-      ) // 8 to 11
-        .toString(16) // in hexadecimal
-    : // or otherwise a concatenated string:
-      (
-        String(1e7) + // 10000000 +
-        String(-1e3) + // -1000 +
-        String(-4e3) + // -4000 +
-        String(-8e3) + // -80000000 +
-        String(-1e11)
-      ) // -100000000000,
-        .replace(
-          // replacing
-          /[018]/g, // zeroes, ones, and eights with
-          UUID, // random hex digits
-        );
-};
-
 export const matchesUrl = (urlArray: string[], urlString: string): boolean => {
   urlString = urlString.replace(/\/$/, '');
 
@@ -87,28 +61,4 @@ export const concatenateQueryParamsOf = (
   });
 
   return resultUrlObj.toString();
-};
-
-export const convertEvaluationVariantToVariant = (
-  evaluationVariant: EvaluationVariant,
-): Variant => {
-  if (!evaluationVariant) {
-    return {};
-  }
-  let experimentKey: string | undefined = undefined;
-  if (evaluationVariant.metadata) {
-    if (typeof evaluationVariant.metadata['experimentKey'] === 'string') {
-      experimentKey = evaluationVariant.metadata['experimentKey'];
-    } else {
-      experimentKey = undefined;
-    }
-  }
-  const variant: Variant = {};
-  if (evaluationVariant.key) variant.key = evaluationVariant.key;
-  if (evaluationVariant.value)
-    variant.value = evaluationVariant.value as string;
-  if (evaluationVariant.payload) variant.payload = evaluationVariant.payload;
-  if (experimentKey) variant.expKey = experimentKey;
-  if (evaluationVariant.metadata) variant.metadata = evaluationVariant.metadata;
-  return variant;
 };
